+++ app/services/mongodb/MongoDBSpaceService.scala
@@ -4,11 +4,11 @@ package services.mongodb
 import javax.inject.{Inject, Singleton}
 import api.Permission
 import api.Permission.Permission
-import com.mongodb.casbah.Imports._
+import com.mongodb.casbah.Imports.{$and, _}
 import com.mongodb.casbah.WriteConcern
 import com.mongodb.casbah.commons.MongoDBObject
 import com.mongodb.DBObject
-import com.novus.salat.dao.{SalatDAO, ModelCompanion}
+import com.novus.salat.dao.{ModelCompanion, SalatDAO}
 import models._
 import org.bson.types.ObjectId
 import play.api.Logger
@@ -675,24 +675,29 @@ class MongoDBSpaceService @Inject() (
    * If entry for this spaceId already exists, adds extractor to it.
    * Otherwise, creates a new entry with spaceId and extractor.
    */
-  def addExtractor (spaceId: UUID, extractor: String) {
-	  //will add extractor to the list of extractors for this space, only if it's not there.
+  def enableExtractor(spaceId: UUID, extractor: String) {
 	  val query = MongoDBObject("spaceId" -> spaceId.stringify)
-	  ExtractorsForSpaceDAO.update(query, $addToSet("extractors" -> extractor), true, false, WriteConcern.Safe)
+	  ExtractorsForSpaceDAO.update(query, $addToSet("enabled" -> extractor), true, false, WriteConcern.Safe)
+  }
+
+  def disableExtractor(spaceId: UUID, extractor: String) {
+    val query = MongoDBObject("spaceId" -> spaceId.stringify)
+    ExtractorsForSpaceDAO.update(query, $addToSet("disabled" -> extractor), true, false, WriteConcern.Safe)
+  }
+
+  def setDefaultExtractor(spaceId: UUID, extractor: String) {
+    val query = MongoDBObject("spaceId" -> spaceId.stringify)
+    ExtractorsForSpaceDAO.update(query, $pull("disabled" -> extractor,"enabled" -> extractor), true, false, WriteConcern.Safe)
   }
 
   /**
    * Returns a list of extractors associated with this spaceId.
    */
-  def getAllExtractors(spaceId: UUID): List[String] = {
-    //Note: in models.ExtractorsForSpace, spaceId must be a String
+  def getAllExtractors(spaceId: UUID): Option[ExtractorsForSpace] = {
+    // Note: in models.ExtractorsForSpace, spaceId must be a String
     // if space Id is UUID, will compile but throws Box run-time error
     val query = MongoDBObject("spaceId" -> spaceId.stringify)
-
-    val list = (for (extr <- ExtractorsForSpaceDAO.find(query)) yield extr).toList
-    //get extractors' names for given space id
-    val extractorList: List[String] = list.flatMap(_.extractors)
-    extractorList
+    ExtractorsForSpaceDAO.findOne(query)
   }